+++ Changelog.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Unreleased
 ### Added
 - Log the versions of our low level libraries.
+- Component `Dropdown`.
 
 ## 4.28.3 - 2021-08-11
 ### Fixed